refactor(provider): extract settings, vision proxy, pricing + request headers
Move the remaining request-path helpers out of extension.ts:
- provider/settings.ts — modelConfigurationSchema / getSettings / modelLimits /
  modelCapabilities / shouldHideDeprecatedModel / resolveRawModelId /
  resolveVendorFromId / getConfiguredApiKey / isVisionProxyEnabled
- provider/visionProxy.ts — proxyVision + showVisionProxyPicker (whole K partition)
- models/pricing.ts — modelPricingFields + costCategory
- request/headers.ts — buildOpenCodeRequestHeaders + stringifyInitiator etc.

extension.ts is now down to ~1400 lines (from 4653). Behavior-preserving;
compile + 291 tests + lint green.
