+++ src/request/headers.ts
@@ -1,0 +1,110 @@
+import * as vscode from "vscode";
+import { OPEN_CODE_CLIENT } from "../config";
+import { getUserAgent } from "../provider/definitions";
+import { messageText } from "../provider/tokens";
+import { isRecord } from "../utils";
+
+// The official OpenCode client sends these headers on every request. The Zen
+// gateway reads x-opencode-session first, then converts that sticky identifier
+// into provider-specific affinity headers such as x-session-affinity upstream.
+//
+// VS Code's provider API does not currently expose a guaranteed public session
+// identifier everywhere, so we first probe a few known internal fields and then
+// fall back to a stable hash of the first messages in the conversation. That
+// preserves sticky routing and cache affinity without depending on hidden state.
+export function buildOpenCodeRequestHeaders(
+  messages: readonly vscode.LanguageModelChatRequestMessage[],
+  options: vscode.ProvideLanguageModelChatResponseOptions,
+  modelId: string,
+): Record<string, string> {
+  const sessionId = cleanHeaderValue(
+    findStringOption(options, [
+      "sessionId",
+      "sessionID",
+      "chatSessionId",
+      "chatSessionID",
+      "conversationId",
+      "conversationID",
+      "threadId",
+      "threadID",
+      "session.id",
+      "chatSession.id",
+    ]) ?? `vscode-${stableHash(conversationAnchor(messages, modelId))}`,
+  );
+  const requestId = cleanHeaderValue(
+    findStringOption(options, ["requestId", "requestID", "messageId", "messageID"]) ??
+      `req-${stableHash(`${String(Date.now())}-${String(Math.random())}-${sessionId}-${modelId}`)}`,
+  );
+
+  return {
+    "x-opencode-session": sessionId,
+    "x-opencode-request": requestId,
+    "x-opencode-client": OPEN_CODE_CLIENT,
+    "User-Agent": getUserAgent(),
+  };
+}
+
+/**
+ * Stringify an arbitrary transport-layer initiator value for diagnostics.
+ * Objects and functions are JSON-serialized, nullish values are dropped, and
+ * primitives are converted directly so logs never show "[object Object]".
+ */
+export function stringifyInitiator(initiator: unknown): string | undefined {
+  if (initiator === undefined || initiator === null) {
+    return undefined;
+  }
+  if (typeof initiator === "string") {
+    return initiator;
+  }
+  if (typeof initiator === "object" || typeof initiator === "function") {
+    return JSON.stringify(initiator);
+  }
+  if (typeof initiator === "symbol" || typeof initiator === "bigint") {
+    return initiator.toString();
+  }
+  if (typeof initiator === "number" || typeof initiator === "boolean") {
+    return String(initiator);
+  }
+  // No known primitive type left; nothing useful to stringify.
+  return undefined;
+}
+
+export function findStringOption(options: unknown, paths: string[]): string | undefined {
+  for (const path of paths) {
+    const value = readPath(options, path.split("."));
+    if (typeof value === "string" && value.trim()) {
+      return value.trim();
+    }
+  }
+  return undefined;
+}
+
+export function readPath(value: unknown, path: string[]): unknown {
+  let current = value;
+  for (const segment of path) {
+    if (!isRecord(current)) {
+      return undefined;
+    }
+    current = current[segment];
+  }
+  return current;
+}
+
+export function conversationAnchor(messages: readonly vscode.LanguageModelChatRequestMessage[], modelId: string): string {
+  const anchorMessages = messages.slice(0, 3).map((message) => `${String(message.role)}:${messageText(message).slice(0, 2048)}`);
+  return anchorMessages.length ? anchorMessages.join("\n") : modelId;
+}
+
+export function cleanHeaderValue(value: string): string {
+  const cleaned = value.replace(/[\r\n]/g, " ").trim();
+  return cleaned ? cleaned.slice(0, 256) : "unknown";
+}
+
+export function stableHash(value: string): string {
+  let hash = 0x811c9dc5;
+  for (let index = 0; index < value.length; index++) {
+    hash ^= value.charCodeAt(index);
+    hash = Math.imul(hash, 0x01000193);
+  }
+  return (hash >>> 0).toString(16).padStart(8, "0");
+}
